backend/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
celery==5.3.4
boto3==1.28.44
stripe==6.7.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.7
redis==5.0.1
gunicorn==21.2.0
pytest==7.4.2
PyJWT==2.8.0 